wallet-service/requirements.txt
fastapi
uvicorn

SQLAlchemy
asyncpg
greenlet

aiosqlite
pydantic-settings

SQLAlchemy
asyncpg
aio_pika